+++ packages/integration-tests/cmr.js
@@ -6,6 +6,7 @@ const xml2js = require('xml2js');
 const { s3 } = require('@cumulus/common/aws');
 const log = require('@cumulus/common/log');
 const { sleep } = require('@cumulus/common/util');
+const { getUrl } = require('@cumulus/cmrjs/utils');
 
 const ONE_SECOND = 1000;
 const THREE_SECONDS = 3000;
@@ -118,24 +119,15 @@ const sampleUmmGranule = {
  * Returns true if the concept exists - if the cmrLink
  * returns a 200 and there are entries
  *
- * @param {string} cmrLink - CMR URL path to concept,
- * i.e. what is returned from post to cmr task
+ * @param {string} cmrLink
+ *   CMR URL path to concept, i.e. what is returned from post to cmr task
+ *   See `@cumulus/cmrjs/cmr-utils/publishECHO10XML2CMR` for expected URL
  * @returns {boolean} true if the concept exists in CMR, false if not
  */
 async function conceptExists(cmrLink) {
-  let response;
-  try {
-    response = await got.get(cmrLink, { json: true });
-  }
-  catch (err) {
-    log.error(err);
-    throw err;
-  }
+  const response = await got.get(cmrLink, { json: true });
 
-  if (response.statusCode !== 200) {
-    log.info(`${cmrLink} conceptExists response: ${response}`);
-    return false;
-  }
+  if (response.statusCode !== 200) return false;
 
   return response.body.feed.entry.length > 0;
 }
@@ -155,11 +147,7 @@ const waitForCmrToBeConsistent = () => sleep(ONE_SECOND);
 async function waitForConceptExistsOutcome(cmrLink, expectation) {
   try {
     await pWaitFor(
-      async () => {
-        const conceptExistsBool = await conceptExists(cmrLink);
-        log.info(`${cmrLink} exists: ${conceptExistsBool}`);
-        return (conceptExistsBool) === expectation;
-      },
+      async () => (await conceptExists(cmrLink)) === expectation,
       { interval: THREE_SECONDS, timeout: ONE_MINUTE }
     );
 
@@ -262,12 +250,12 @@ async function generateAndStoreCmrXml(granule, collection, bucket, additionalUrl
 }
 
 /**
- * Transforms a cmrfiletype to a version string or returns an empty string.
+ * Transforms a CMR metadata format to a version string or returns an empty string.
  *
  * @param {string} typeStr
- * @returns {string} the decoded version or empty strign if a version can't be created.
+ * @returns {string} the decoded version or empty string if a version can't be created.
  */
-function fileTypeToVersion(typeStr) {
+function metadataFormatToVersion(typeStr) {
   try {
     return typeStr.match(/umm_json_v(.*)/)[1].replace('_', '.');
   }
@@ -277,12 +265,83 @@ function fileTypeToVersion(typeStr) {
 }
 
 /**
- * tester to determine if the input cmrFiletype is a UMM JSON file.
- * @param {string} cmrFileType
- * @returns {boolean} true if the cmrFiletype matches umm_json_v
+ * tester to determine if the input cmrMetadataFormat is a UMM JSON file.
+ * @param {string} cmrMetadataFormat
+ * @returns {boolean} true if the cmrMetadataFormat matches umm_json_v
  */
-function isUMMGFileType(cmrFileType) {
-  return cmrFileType && cmrFileType.match(/umm_json_v/);
+function isUMMGMetadataFormat(cmrMetadataFormat) {
+  return cmrMetadataFormat && cmrMetadataFormat.match(/umm_json_v/);
+}
+
+/**
+ * Get the online resource links from the CMR objects for ECH010
+ *
+ * @param {string} cmrLink
+ *   CMR URL path to concept, i.e. what is returned from post to cmr task
+ * @returns {Array<Object>} Array of link objects in the format
+ * { inherited: true,
+    rel: 'http://esipfed.org/ns/fedsearch/1.1/metadata#',
+    hreflang: 'en-US',
+    href: 'https://opendap.cr.usgs.gov/opendap/hyrax/MYD13Q1.006/contents.html' }
+ */
+async function getOnlineResourcesECHO10(cmrLink) {
+  const response = await got.get(cmrLink);
+
+  if (response.statusCode !== 200) {
+    return null;
+  }
+
+  const body = JSON.parse(response.body);
+
+  const links = body.feed.entry.map((e) => e.links);
+
+  // Links is a list of a list, so flatten to be one list
+  return [].concat(...links);
+}
+
+/**
+ * Get the online resource links from the CMR objects for UMM-G
+ *
+ * @param {string} cmrLink
+ *   CMR URL path to concept, i.e. what is returned from post to cmr task
+ * @returns {Array<Object>} Array of link objects in the format
+ * { URL: "https://example.com/cumulus-test-sandbox-protected/MOD09GQ___006/2016/MOD/MOD09GQ.A0794505._4kqJd.006.9457902462263.hdf",
+    Description: "File to download",
+    Type: "GET DATA" }
+ */
+async function getOnlineResourcesUMMG(cmrLink) {
+  const response = await got.get(cmrLink);
+
+  if (response.statusCode !== 200) {
+    return null;
+  }
+
+  const body = JSON.parse(response.body);
+
+  const links = body.items.map((item) => item.umm.RelatedUrls);
+
+  // Links is a list of a list, so flatten to be one list
+  return [].concat(...links);
+}
+
+/**
+ * Fetches online resources from CMR based on file type (ECHO10, UMM-G)
+ *
+ * @param {Object} granule
+ * @param {string} granule.cmrMetadataFormat - the cmr file type (e.g. echo10, umm-g)
+ * @param {Object} granule.cmrConceptId - the CMR granule concept ID
+ * @param {Object} granule.cmrLink - the metadata's granuleId
+ *
+ * @returns {Promise<Array<Object>>} - Promise returning array of links
+ */
+async function getOnlineResources({ cmrMetadataFormat, cmrConceptId, cmrLink }) {
+  if (cmrMetadataFormat === 'echo10') {
+    return getOnlineResourcesECHO10(cmrLink);
+  }
+  if (isUMMGMetadataFormat(cmrMetadataFormat)) {
+    return getOnlineResourcesUMMG(`${getUrl('search')}granules.umm_json?concept_id=${cmrConceptId}`);
+  }
+  throw new Error(`Invalid cmrMetadataFormat passed to getOnlineResources: ${cmrMetadataFormat}}`);
 }
 
 /**
@@ -293,7 +352,7 @@ function isUMMGFileType(cmrFileType) {
  * @param {Object} collection - collection object
  * @param {string} bucket - bucket to save the xml file to
  * @param {Array<string>} additionalUrls - URLs to convert to related urls
- * @param {string} cmrFileType - CMR UMM-G version string <umm_json_v[x.y]>
+ * @param {string} cmrMetadataFormat - CMR UMM-G version string <umm_json_v[x.y]>
  * @returns {Promise<Array<string>>} - Promise of a list of granule files including the created
  * CMR files
  */
@@ -302,9 +361,9 @@ async function generateAndStoreCmrUmmJson(
   collection,
   bucket,
   additionalUrls,
-  cmrFileType
+  cmrMetadataFormat
 ) {
-  const versionString = fileTypeToVersion(cmrFileType);
+  const versionString = metadataFormatToVersion(cmrMetadataFormat);
   const jsonObject = sampleUmmGranule;
   jsonObject.GranuleUR = granule.granuleId;
 
@@ -357,7 +416,7 @@ async function generateAndStoreCmrUmmJson(
  *                                   output
  * @param {Object} collection - collection object that includes name and version
  * @param {string} bucket - location to save the xmls to
- * @param {string} cmrFileType - CMR file type to generate. Options are echo10, umm_json_v1_4,
+ * @param {string} cmrMetadataFormat - CMR file type to generate. Options are echo10, umm_json_v1_4,
  *                               umm_json_v1_5, (and likely umm_json_v1_<x>). The default is echo10
  * @param {Array<string>} additionalUrls - URLs to convert to online resources or related urls
  * @returns {Array<string>} list of S3 locations for CMR xml files
@@ -366,17 +425,17 @@ async function generateCmrFilesForGranules(
   granules,
   collection,
   bucket,
-  cmrFileType,
+  cmrMetadataFormat,
   additionalUrls
 ) {
   let files;
 
-  log.info(`Generating fake CMR file with type ${cmrFileType}`);
+  log.info(`Generating fake CMR file with type ${cmrMetadataFormat}`);
 
-  if (isUMMGFileType(cmrFileType)) {
+  if (isUMMGMetadataFormat(cmrMetadataFormat)) {
     files = await Promise.all(
       granules.map(
-        (g) => generateAndStoreCmrUmmJson(g, collection, bucket, additionalUrls, cmrFileType)
+        (g) => generateAndStoreCmrUmmJson(g, collection, bucket, additionalUrls, cmrMetadataFormat)
       )
     );
   }