+++ CHANGELOG.md
@@ -6,7 +6,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+**Please Note**
+- If you are using the `@cumulus/post-to-cmr` task in your workflow and specifying the metadata format to post using the `cmrFileType` config property, you must update that workflow step to use the `cmrMetadataFormat` config property instead and re-deploy
+
 ### Added
+
+- **CUMULUS-672**
+  - Added `cmrMetadataFormat` and `cmrConceptId` to output for individual granules from `@cumulus/post-to-cmr`. `cmrMetadataFormat` will default to value from workflow configuration, if provided, otherwise it will attempt to read the `cmrMetadataFormat` generated in `@cumulus/cmrjs/publish2CMR`
+  - Added helpers to `@packages/integration-tests/api/distribution`:
+    - `getDistributionApiFileStream` returns a stream to download files protected by the distribution API
+    - `getDistributionFileUrl` constructs URLs for requesting files from the distribution API
+
 - CUMULUS-1171
   - Added `@cumulus/common` API documentation to `packages/common/docs/API.md`
   - Added an `npm run build-docs` task to `@cumulus/common`
@@ -24,6 +34,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - CUMULUS-1151 Compare the granule holdings in CMR with Cumulus' internal data store
 
 ### Changed
+
+- **CUMULUS-672**
+  - `@cumulus/post-to-cmr` now looks in `config.cmrMetadataFormat` instead of `config.cmrFileType` to determine the type of metadata to post to CMR
+  - Changed `@cumulus/integration-tests/api/EarthdataLogin.getEarthdataLoginRedirectResponse` to `@cumulus/integration-tests/api/EarthdataLogin.getEarthdataAccessToken`. The new function returns an access response from Earthdata login, if successful.
+  - `@cumulus/integration-tests/cmr/getOnlineResources` now accepts an object of options, including `cmrMetadataFormat`. Based on the `cmrMetadataFormat`, the function will correctly retrieve the online resources for each metadata format (ECHO10, UMM-G)
+
 - CUMULUS-1171
   - NOTE: This is a breaking change. When applying this upgrade, users will need to:
     1. Disable all workflow rules